integrated front-end basic structure

+++ src/main/java/ebi/uk/eva/controller/ProjectController.java
@@ -2,8 +2,9 @@ package ebi.uk.eva.controller;
 
 import java.util.List;
 
-import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -21,34 +22,53 @@ import ebi.uk.eva.service.ProjectService;
 @RestController
 @RequestMapping("/projects")
 public class ProjectController {
-	
-	final static Logger logger = Logger.getLogger(ProjectController.class);
 
 	@Autowired
 	private ProjectService projectService;
 
 	@RequestMapping(method = RequestMethod.GET)
-	public List<Project> getAllProjects() {
-		return projectService.getAllProjects();
+	public ResponseEntity<List<Project>> getAllProjects() {
+		List<Project> projects = projectService.getAllProjects();
+		if (projects.isEmpty()) {
+			return new ResponseEntity<List<Project>>(HttpStatus.NO_CONTENT);
+		}
+		return new ResponseEntity<List<Project>>(projects, HttpStatus.OK); 
 	}
 
 	@RequestMapping(method = RequestMethod.GET, value = "/{id}")
-	public Project getProject(@PathVariable String id) {
-		return projectService.getProject(id);
+	public ResponseEntity<Project> getProject(@PathVariable String id) {
+		Project project = projectService.getProject(id);
+		if (project == null) {
+			return new ResponseEntity<Project>(HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<Project>(project, HttpStatus.OK); 
 	}
 
 	@RequestMapping(method = RequestMethod.POST)
-	public void addProject(@RequestBody Project project) {
+	public ResponseEntity<Project> addProject(@RequestBody Project project) {
 		projectService.addProject(project);
+		return new ResponseEntity<Project>(project, HttpStatus.CREATED);
 	}
 
 	@RequestMapping(method = RequestMethod.PUT, value = "/{id}")
-	public void addProject(@RequestBody Project project, @PathVariable String id) {
-		projectService.updateProject(project, id);
+	public ResponseEntity<Void> addProject(@RequestBody Project project, @PathVariable String id) {
+		Project foundProject = projectService.getProject(id);
+		if (foundProject == null) {
+			return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+		} else {
+			projectService.updateProject(project, id);
+			return new ResponseEntity<Void>(HttpStatus.OK);
+		}
 	}
 
 	@RequestMapping(method = RequestMethod.DELETE, value = "/{id}")
-	public void deleteProject(@PathVariable String id) {
-		projectService.deleteProject(id);
+	public ResponseEntity<Void> deleteProject(@PathVariable String id) {
+		Project foundProject = projectService.getProject(id);
+		if (foundProject == null) {
+			return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+		} else {
+			projectService.deleteProject(id);
+			return new ResponseEntity<Void>(HttpStatus.GONE);
+		}
 	}
 }
